refactor: 'Intro' view to use new 'AbstractView'
